Update tutorial-discover-physical.md
This was notated when troubleshooting with customer, the 'Set up prerequisites' check succeeded, but when attempting to Register with migrate failed unless proxy was set within web page form. Setting in the browser settings and system level did not work.
My suggested changes is to make this step more clear, feel free to re-word if necessary.